requirements.txt
markdown
PyYAML
requests
beautifulsoup4
lxml
PyQt5
PyQtWebEngine
Pillow
